% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{binningY_d1}
\alias{binningY_d1}
\title{Binning on Y}
\usage{
binningY_d1(X, Y, B = 40, min_count_ratio = 0)
}
\arguments{
\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{Y}{length T list with \code{Y[[t]]} being a n_t-by-d matrix}

\item{min_count_ratio}{min count ratio}

\item{number}{of binning}
}
\value{
Y_bin N_bin-by-d matrix, indicating the center of the bins

Y_count N_bin-vector indicating how many points belong to each bin

X_bin N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's

time_indicator length N_bin factor indicating the time of the data points

N the total number of observations of Y_i^(t) in Y
}
\description{
Binning on Y
}