% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{calc_resi}
\alias{calc_resi}
\title{calculating residuals for each k}
\usage{
calc_resi(X_bin, Y_bin, theta0, theta, K)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{Y_bin}{N_bin-by-d matrix, indicating the center of the bin}

\item{theta0}{length K list, with \code{theta0[[k]]} being the estimate for the intercept coefficient
of the regression for kth group}

\item{theta}{length K list, with \code{theta[[k]]} being the p-by-1 vector. Initial estimate for coefficients
of the regression for kth group}

\item{K}{number of clusters}
}
\value{
length K list, with \code{resi0[[k]]} N_bin-by-d residual matrix
}
\description{
calculating residuals for each k
}