% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{calc_surr}
\alias{calc_surr}
\title{calculating the surrogate loglikelihood Q}
\usage{
calc_surr(
  X_bin,
  g,
  resi,
  theta,
  alpha,
  weight,
  idx,
  lambda_alpha,
  lambda_theta,
  K,
  N
)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{g}{length K list, with \code{g[[k]]} being \code{mlelcd} for the kth group}

\item{resi}{length K list, with \code{resi[[k]]} N_bin-by-d residual matrix}

\item{theta}{length K list, with \code{theta[[k]]} being the p-by-1 vector. Initial estimate for coefficients
of the regression for kth group}

\item{alpha}{(p+1)-by-K matrix. The coefficients for the cluster probabilities.}

\item{weight}{N_bin by K matrix. weight\link{i, k} represents weight of ith residual for kth group}

\item{idx}{N_bin by K logical matrix. idx\link{i, k} represents whether the corresponding weight is above r_bar}

\item{lambda_alpha}{penalty parameter for alpha}

\item{lambda_theta}{penalty parameter for theta}

\item{K}{number of clusters}

\item{N}{the total number of observations of Y_i^(t) in Y}
}
\value{
Q the surrogate loglikelihood of the current parameters
}
\description{
calculating the surrogate loglikelihood Q
}