% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{E_step}
\alias{E_step}
\title{Updating responsibility}
\usage{
E_step(X_bin, Y_count, resi, alpha, g, K, r_bar)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{Y_count}{N_bin-vector indicating how many points belong to each bin}

\item{resi}{length K list, with \code{resi[[k]]} N_bin-by-d residual matrix}

\item{alpha}{(p+1)-by-K matrix. The coefficients for the cluster probabilities.}

\item{g}{length K list, with \code{g[[k]]} being \code{mlelcd} for the kth group}

\item{K}{number of clusters}

\item{r_bar}{threshold for responsibility}
}
\value{
resp N_bin by K matrix. \code{resp[i, k]} being the current estimate for P(Z_i = k|X, Y)

weight N_bin by K matrix. weight\link{i, k} represents weight of ith residual for kth group

idx N_bin by K logical matrix. idx\link{i, k} represents whether the corresponding weight is above r_bar
}
\description{
Updating responsibility
}