% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{generate_mix_lcd_reg}
\alias{generate_mix_lcd_reg}
\title{Generate data from mixture of log-concave regressions}
\usage{
generate_mix_lcd_reg(
  X,
  theta0,
  theta,
  alpha,
  num_points,
  err_type = c("Gaussian", "exp")
)
}
\arguments{
\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{theta0}{length K list, with \code{theta0[[k]]} being the estimate for the intercept coefficient
of the regression for kth group}

\item{theta}{length K list, with \code{theta[[k]]} being the p-by-1 vector. Initial estimate for coefficients
of the regression for kth group}

\item{alpha}{(p+1)-by-K matrix. The coefficients for the cluster probabilities.}

\item{num_points}{T-vector of integers giving the number of points n_t to
generate at each time point t.}

\item{err_type}{a K-vector of strings, each of which is the type of error distribution
for the each group. Now, it is one of 'Gaussian', 'exp'}
}
\value{
Y length T list with \code{Y[[t]]} being a n_t-by-d matrix

Z length T list with \code{Z[[t]]} being a n_t vector of cluster memberships
}
\description{
The mixture of log-concave regressions model is defined as follows:
}
\details{
At time t there are n_t points generated as follows:

Given Z_i^{(t)} = k, Y_i^{(t)}  = theta_{k0} + theta_k^T X^{(t)} + epsilon_i^{(t)}

where epsilon_i^{(t)} ~ exp(g_k)

and P(Z_i^{(t)} =k |X^{(t)} ;alpha) = \frac{exp(alpha_{k0} + alpha_k^T X^{(t)} )}{\sum_{l=1}^K exp(alpha_{l0} + alpha_l^T X^{(t)} )}

This function generates Y and Z

For simplicity we here assume alpha = 0, K = 2, p = 2 and d = 1
}