% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{LP_d1}
\alias{LP_d1}
\title{Updating theta for each k}
\usage{
LP_d1(
  X_bin,
  X,
  Y_bin,
  g_k,
  weight_k,
  idx_old_k,
  theta0_k,
  theta_k,
  lambda_theta,
  time_indicator
)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{Y_bin}{N_bin-by-d matrix, indicating the center of the bin}

\item{g_k}{\code{mlelcd} for the kth group}

\item{weight_k}{N_bin vector of weight for kth group}

\item{idx_old_k}{N_bin logical vector, but used in the previous iteration.
idx_old_k\link{i} represents whether the corresponding previous weight for kth group is above r_bar}

\item{theta0_k}{an intercept coefficient for the regression for the kth group}

\item{theta_k}{a p-vector of coefficients for the regression for the kth group}

\item{lambda_theta}{penalty parameter for theta}

\item{time_indicator}{length N_bin factor indicating the time of the data points}
}
\value{
theta0_k estimate for the intercept coefficient of the regression for kth group

theta_k p-vector of estimates for coefficients of the regression for kth group
}
\description{
Updating theta for each k
}