% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{mixLcdReg-package}
\alias{mixLcdReg}
\alias{mixLcdReg-package}
\title{Mixture of log-concave regressions modeling}
\description{
This package uses mixture of log-concave regressions to analyze complex and heterogenous data.
}
\author{
\strong{Maintainer}: Sang-wook Lee \email{sangwook@usc.edu}

}
\keyword{internal}