% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_alpha}
\alias{Mstep_alpha}
\title{updating alpha}
\usage{
Mstep_alpha(X, weight, lambda_alpha, time_indicator)
}
\arguments{
\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{weight}{N_bin by K matrix. weight\link{i, k} represents weight of ith residual for kth group}

\item{lambda_alpha}{penalty parameter for alpha}

\item{time_indicator}{length N_bin factor indicating the time of the data points}
}
\value{
alpha (p+1)-by-K matrix. The coefficients for the cluster probabilities.
}
\description{
updating alpha
}