% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_g}
\alias{Mstep_g}
\title{updating \code{mlelcd} g's}
\usage{
Mstep_g(resi, weight, idx, K)
}
\arguments{
\item{resi}{length K list, with \code{resi[[k]]} N_bin-by-d residual matrix}

\item{weight}{N_bin by K matrix. weight\link{i, k} represents weight of ith residual for kth group}

\item{idx}{N_bin by K logical matrix. idx\link{i, k} represents whether the corresponding weight is above r_bar}

\item{K}{number of clusters}
}
\value{
g length K list, with \code{g[[k]]} being \code{mlelcd} for the kth group
}
\description{
updating \code{mlelcd} g's
}