% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_theta}
\alias{Mstep_theta}
\title{Updating theta}
\usage{
Mstep_theta(
  X_bin,
  X,
  Y_bin,
  g,
  weight,
  idx_old,
  theta0,
  theta,
  lambda_theta,
  K,
  time_indicator
)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{Y_bin}{N_bin-by-d matrix, indicating the center of the bin}

\item{g}{length K list, with \code{g[[k]]} being \code{mlelcd} for the kth group}

\item{weight}{N_bin by K matrix. weight\link{i, k} represents weight of ith residual for kth group}

\item{idx_old}{N_bin by K logical matrix, but used in the previous iteration.
idx\link{i, k} represents whether the corresponding previous weight is above r_bar}

\item{theta0}{length K list, with \code{theta0[[k]]} being the estimate for the intercept coefficient
of the regression for kth group}

\item{theta}{length K list, with \code{theta[[k]]} being the p-by-1 vector. Initial estimate for coefficients
of the regression for kth group}

\item{lambda_theta}{penalty parameter for theta}

\item{K}{number of clusters}

\item{time_indicator}{length N_bin factor indicating the time of the data points}
}
\value{
theta0 length K list, with \code{theta0[[k]]} being the estimate for the intercept coefficient
of the regression for kth group

theta length K list, with \code{theta[[k]]} being the p-by-1 vector of estimate for coefficients
of the regression for kth group
}
\description{
Updating theta
}