% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{pi_k}
\alias{pi_k}
\title{calculating \pi_{tk}(\alpha) = P(Z_i^{(t)} = k| X^(t))}
\usage{
pi_k(X_bin, alpha)
}
\arguments{
\item{X_bin}{N_bin-by-p matrix, which is expanded version of X so that its dimension agrees with Y_bin's}

\item{alpha}{(p+1)-by-K matrix. The coefficients for the cluster probabilities.}
}
\value{
pi_k N_bin-by-K matrix with \code{pi_k[i,k]} indicates P(Z_i = k| X^(t))
}
\description{
calculating \pi_{tk}(\alpha) = P(Z_i^{(t)} = k| X^(t))
}