% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_data}
\alias{plot_data}
\title{Plot raw data when \code{d = 1}}
\usage{
plot_data(Y)
}
\arguments{
\item{Y}{length T list with \code{Y[[t]]} being a n_t-by-d matrix}
}
\description{
Plot raw data when \code{d = 1}
}