% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_data_and_model}
\alias{plot_data_and_model}
\title{Plot data colored by cluster assignment with cluster means when \code{d=1}}
\usage{
plot_data_and_model(X, Y, Z, theta0, theta)
}
\arguments{
\item{X}{a T-by-p matrix of covariates, where \code{X[[t]]} being the p-vector of independent variable at time t}

\item{Y}{length T list with \code{Y[[t]]} being a n_t-by-d matrix}

\item{Z}{a length T list with \code{Z[[t]]} being a n_t vector of cluster assignments}

\item{theta0}{length K list, with \code{theta0[[k]]} being the estimate for the intercept coefficient
of the regression for kth group}

\item{theta}{length K list, with \code{theta[[k]]} being the p-by-1 vector. Initial estimate for coefficients
of the regression for kth group}
}
\description{
Plot data colored by cluster assignment with cluster means when \code{d=1}
}